Performance improvement
Internal ip db now has a state column to keep track of state for firewall. This allows passing a delta enumerator to the firewall that will only remove ips from rules that have them, and then add ips to any rule with space in it. For Linux this does not change things much, but for the Windows firewall with a 1000 ip max per rule, it helps performance quite a bit.