Tiered ban times
BanTime allows a comma separated array of time spans. If there are multiple ban times, each unban will leave the ip address in the db and move to the next time span when the ip address has enough failed logins.

The failed login count can be reset to 0 or left as is. If reset to 0, then the failed logins will increment until reaching the threshold, otherwise if it is left as is, the very next failed login will cause an immediate ban.