Allow capturing relevant log snippet
Using a new log group, the relevant parts of the log can be captured. The log event object has a LogData property now which contains this log group value, if any.

In the future, this LogData can be included in the ipban logs or be sent on to ipthreat or other providers to provider further context on why the ip address was reported.